requirements.txt
pdfplumber
python-docx
nltk
scikit-learn
streamlit